Rework the fix
